Changed MessageDispatcher.dispatch to return byte flags. Extracted interface PartitionIndex. Added CompactedTopicMessageDispatcher.
