+++ src/main/java/org/reaktivity/nukleus/kafka/internal/stream/CompactedTopicMessageDispatcher.java
@@ -1,0 +1,68 @@
+/**
+ * Copyright 2016-2017 The Reaktivity Project
+ *
+ * The Reaktivity Project licenses this file to you under the Apache License,
+ * version 2.0 (the "License"); you may not use this file except in compliance
+ * with the License. You may obtain a copy of the License at:
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations
+ * under the License.
+ */
+package org.reaktivity.nukleus.kafka.internal.stream;
+
+import java.util.Iterator;
+
+import org.agrona.DirectBuffer;
+import org.reaktivity.nukleus.kafka.internal.cache.IPartitionIndex;
+import org.reaktivity.nukleus.kafka.internal.cache.IPartitionIndex.Entry;
+import org.reaktivity.nukleus.kafka.internal.types.OctetsFW;
+
+public class CompactedTopicMessageDispatcher extends TopicMessageDispatcher
+{
+    private final IPartitionIndex[] indexes;
+
+    public CompactedTopicMessageDispatcher(
+        IPartitionIndex[] indexes)
+    {
+        super(indexes.length, KeyMessageDispatcher::new, CompactedHeaderValueMessageDispatcher::new);
+        this.indexes = indexes;
+    }
+
+    public byte dispatch(
+                 int partition,
+                 long requestOffset,
+                 long messageOffset,
+                 DirectBuffer key,
+                 HeadersFW headers,
+                 long timestamp,
+                 long traceId,
+                 DirectBuffer value)
+    {
+        byte result = super.dispatch(partition, requestOffset, messageOffset, key, headers::supplyHeader, timestamp,
+                traceId, value);
+        indexes[partition].add(requestOffset, messageOffset, timestamp, traceId, key, headers, value);
+        return result;
+    }
+
+    @Override
+    public Entry getEntry(
+        int partition,
+        OctetsFW key)
+    {
+        return indexes[partition].getEntry(partition, key);
+    }
+
+    @Override
+    public Iterator<Entry> entries(
+        int partition,
+        long requestOffset)
+    {
+        return indexes[partition].entries(requestOffset);
+    }
+
+}
