+++ src/main/java/org/reaktivity/nukleus/kafka/internal/stream/FetchResponseDecoder.java
@@ -70,7 +70,7 @@ public class FetchResponseDecoder implements ResponseDecoder
     private final UnsafeBuffer keyBuffer = new UnsafeBuffer(BufferUtil.EMPTY_BYTE_ARRAY);
     private final UnsafeBuffer valueBuffer = new UnsafeBuffer(BufferUtil.EMPTY_BYTE_ARRAY);
 
-    private final Function<String, MessageDispatcher> getDispatcher;
+    private final Function<String, DecoderMessageDispatcher> getDispatcher;
     private final StringIntToLongFunction getRequestedOffsetForPartition;
     private final StringIntShortConsumer errorHandler;
     private final int maxRecordBatchSize;
@@ -84,7 +84,7 @@ public class FetchResponseDecoder implements ResponseDecoder
 
     private int topicCount;
     private String topicName;
-    private MessageDispatcher topicDispatcher;
+    private DecoderMessageDispatcher messageDispatcher;
     private int partitionCount;
     private int recordSetBytesRemaining;
     private int recordBatchBytesRemaining;
@@ -103,7 +103,7 @@ public class FetchResponseDecoder implements ResponseDecoder
     private final BufferBytesDecoderState bufferBytesDecoderState = new BufferBytesDecoderState();
 
     FetchResponseDecoder(
-        Function<String, MessageDispatcher> getDispatcher,
+        Function<String, DecoderMessageDispatcher> getDispatcher,
         StringIntToLongFunction getRequestedOffsetForPartition,
         StringIntShortConsumer errorHandler,
         MutableDirectBuffer decodingBuffer,
@@ -273,7 +273,7 @@ public class FetchResponseDecoder implements ResponseDecoder
             {
                 topicCount--;
                 topicName = response.name().asString();
-                topicDispatcher = getDispatcher.apply(topicName);
+                messageDispatcher = getDispatcher.apply(topicName);
                 partitionCount = response.partitionCount();
                 decoderState = this::decodePartitionResponse;
                 newOffset = response.limit();
@@ -364,7 +364,7 @@ public class FetchResponseDecoder implements ResponseDecoder
                 if (highWatermark > requestedOffset)
                 {
                     nextFetchAt = requestedOffset + 1;
-                    topicDispatcher.flush(partition, requestedOffset, requestedOffset + 1);
+                    messageDispatcher.flush(partition, requestedOffset, requestedOffset + 1);
                 }
                 decoderState = this::decodePartitionResponse;
             }
@@ -374,7 +374,7 @@ public class FetchResponseDecoder implements ResponseDecoder
                         "[nukleus-kafka] skipping topic: %s partition: %d offset: %d because record set size %d " +
                         "exceeds configured partition.max.bytes %d",
                         topicName, partition, requestedOffset, recordSetBytesRemaining, maxRecordBatchSize);
-                topicDispatcher.flush(partition, requestedOffset, requestedOffset + 1);
+                messageDispatcher.flush(partition, requestedOffset, requestedOffset + 1);
                 skipBytesDecoderState.bytesToSkip = recordSetBytesRemaining;
                 skipBytesDecoderState.nextState = this::decodePartitionResponse;
                 decoderState = skipBytesDecoderState;
@@ -405,7 +405,7 @@ public class FetchResponseDecoder implements ResponseDecoder
             // Truncated record batch at end of record set
             if (nextFetchAt > requestedOffset)
             {
-                topicDispatcher.flush(partition, requestedOffset, nextFetchAt);
+                messageDispatcher.flush(partition, requestedOffset, nextFetchAt);
             }
             skipBytesDecoderState.bytesToSkip = recordSetBytesRemaining;
             skipBytesDecoderState.nextState = this::decodePartitionResponse;
@@ -424,7 +424,7 @@ public class FetchResponseDecoder implements ResponseDecoder
                         "[nukleus-kafka] skipping large message at topic: %s partition: %d offset: %d batch-size: %d\n",
                         topicName, partition, recordBatch.firstOffset(), recordBatch.length());
                     nextFetchAt = recordBatch.firstOffset() + 1;
-                    topicDispatcher.flush(partition, requestedOffset, nextFetchAt);
+                    messageDispatcher.flush(partition, requestedOffset, nextFetchAt);
                     skipBytesDecoderState.bytesToSkip = recordSetBytesRemaining;
                     skipBytesDecoderState.nextState = this::decodePartitionResponse;
                     decoderState = skipBytesDecoderState;
@@ -467,7 +467,7 @@ public class FetchResponseDecoder implements ResponseDecoder
                 // offset of the last record in the batch. So we must use this to make sure we continue to advance.
                 nextFetchAt = firstOffset + lastOffsetDelta + 1;
             }
-            topicDispatcher.flush(partition, requestedOffset, nextFetchAt);
+            messageDispatcher.flush(partition, requestedOffset, nextFetchAt);
             decoderState = this::decodePartitionResponse;
         }
         else if (recordBatchBytesRemaining == 0)
@@ -485,7 +485,7 @@ public class FetchResponseDecoder implements ResponseDecoder
                 {
                     // Truncated record at end of batch, make sure we attempt to re-fetch it
                     nextFetchAt = firstOffset + lastOffsetDelta;
-                    topicDispatcher.flush(partition, requestedOffset, nextFetchAt);
+                    messageDispatcher.flush(partition, requestedOffset, nextFetchAt);
                     skipBytesDecoderState.bytesToSkip = recordSetBytesRemaining;
                     skipBytesDecoderState.nextState = this::decodePartitionResponse;
                     decoderState = skipBytesDecoderState;
@@ -564,8 +564,8 @@ public class FetchResponseDecoder implements ResponseDecoder
                 else
                 {
                     headers.wrap(buffer, headersOffset, headersLimit);
-                    topicDispatcher.dispatch(partition, requestedOffset, nextFetchAt,
-                            key, headers::supplyHeader, timestamp, traceId, value);
+                    messageDispatcher.dispatch(partition, requestedOffset, nextFetchAt,
+                            key, headers, timestamp, traceId, value);
                 }
             }
             newOffset = headersLimit;
