+++ src/main/java/org/reaktivity/nukleus/kafka/internal/stream/HeadersMessageDispatcher.java
@@ -42,7 +42,7 @@ public class HeadersMessageDispatcher implements MessageDispatcher
     }
 
     @Override
-    public int dispatch(
+    public byte dispatch(
                  int partition,
                  long requestOffset,
                  long messageOffset,
@@ -52,12 +52,12 @@ public class HeadersMessageDispatcher implements MessageDispatcher
                  long traceId,
                  DirectBuffer value)
     {
-        int result = 0;
-        result +=  broadcast.dispatch(partition, requestOffset, messageOffset, key, supplyHeader, timestamp, traceId, value);
+        byte result = 0;
+        result |=  broadcast.dispatch(partition, requestOffset, messageOffset, key, supplyHeader, timestamp, traceId, value);
         for (int i = 0; i < dispatchers.size(); i++)
         {
             MessageDispatcher dispatcher = dispatchers.get(i);
-            result += dispatcher.dispatch(partition, requestOffset, messageOffset, key, supplyHeader, timestamp, traceId, value);
+            result |= dispatcher.dispatch(partition, requestOffset, messageOffset, key, supplyHeader, timestamp, traceId, value);
         }
         return result;
     }
