+++ src/test/java/org/reaktivity/nukleus/kafka/internal/stream/BroadcastMessageDispatcherTest.java
@@ -19,6 +19,8 @@ import static java.nio.charset.StandardCharsets.UTF_8;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
+import static org.reaktivity.nukleus.kafka.internal.stream.MessageDispatcher.FLAGS_DELIVERED;
+import static org.reaktivity.nukleus.kafka.internal.stream.MessageDispatcher.FLAGS_MATCHED;
 
 import java.util.function.Function;
 
@@ -64,13 +66,14 @@ public final class BroadcastMessageDispatcherTest
             {
                 oneOf(child1).dispatch(with(1), with(10L), with(12L), with(bufferMatching("key")),
                         with(header), with(timestamp), with(traceId), with((DirectBuffer) null));
-                will(returnValue(1));
+                will(returnValue(FLAGS_DELIVERED));
                 oneOf(child2).dispatch(with(1), with(10L), with(12L), with(bufferMatching("key")),
                         with(header), with(timestamp), with(traceId), with((DirectBuffer) null));
-                will(returnValue(1));
+                will(returnValue(FLAGS_MATCHED));
             }
         });
-        assertEquals(2, dispatcher.dispatch(1, 10L, 12L, asBuffer("key"), header, timestamp, traceId, null));
+        assertEquals(FLAGS_DELIVERED | FLAGS_MATCHED,
+                dispatcher.dispatch(1, 10L, 12L, asBuffer("key"), header, timestamp, traceId, null));
     }
 
     @Test
